+++ packages/expo-router/src/react-navigation/stack/types.tsx
@@ -47,6 +47,22 @@ export type StackNavigationEventMap = {
   gestureCancel: { data: undefined };
 };
 
+export type StackViewEmit = (
+  event:
+    | {
+        type: 'transitionStart' | 'transitionEnd';
+        target?: string;
+        data: { closing: boolean };
+      }
+    | {
+        type: 'gestureStart' | 'gestureEnd' | 'gestureCancel';
+        target?: string;
+        data?: undefined;
+      }
+) => void;
+
+export type StackViewState = Pick<StackNavigationState<ParamListBase>, 'index' | 'routes'>;
+
 export type StackNavigationHelpers = NavigationHelpers<ParamListBase, StackNavigationEventMap> &
   StackActionHelpers<ParamListBase>;